#1273-moved-utils
