+++ src/backend/src/utils/change-requests.utils.ts
@@ -1,7 +1,6 @@
 import prisma from '../prisma/prisma';
-import { Scope_CR_Why_Type, Team, User, Prisma, Change_Request, Change, WBS_Element } from '@prisma/client';
+import { Scope_CR_Why_Type, User, Prisma, Change_Request, Change } from '@prisma/client';
 import { addWeeksToDate, ChangeRequestReason } from 'shared';
-import { reactToMessage, replyToMessageInThread, sendMessage } from '../integrations/slack';
 import { HttpException, NotFoundException } from './errors.utils';
 import { ChangeRequestStatus } from 'shared';
 import changeRequestRelationArgs from '../prisma-query-args/change-requests.query-args';
@@ -21,146 +20,6 @@ export const convertCRScopeWhyType = (whyType: Scope_CR_Why_Type): ChangeRequest
     OTHER_PROJECT: ChangeRequestReason.OtherProject,
     OTHER: ChangeRequestReason.Other
   }[whyType]);
-
-/**
- * Sends slack notifications to teams for new CRs and returns the messages sent in slack
- *
- * @param team the teams of the cr to notify
- * @param message the message to send to the teams
- * @param crId the cr id
- * @param budgetImpact the amount of budget requested for the cr
- * @returns the channelId and timestamp of the messages sent in slack
- */
-export const sendSlackChangeRequestNotification = async (
-  team: Team,
-  message: string,
-  crId: number,
-  budgetImpact?: number
-): Promise<{ channelId: string; ts: string }[]> => {
-  if (process.env.NODE_ENV !== 'production') return []; // don't send msgs unless in prod
-  const msgs: { channelId: string; ts: string }[] = [];
-  const fullMsg = `:tada: New Change Request! :tada: ${message}`;
-  const fullLink = `https://finishlinebyner.com/cr/${crId}`;
-  const btnText = `View CR #${crId}`;
-  const notification = await sendMessage(team.slackId, fullMsg, fullLink, btnText);
-  if (notification) msgs.push(notification);
-
-  if (budgetImpact && budgetImpact > 100) {
-    const importantNotification = await sendMessage(
-      process.env.SLACK_EBOARD_CHANNEL!,
-      `${fullMsg} with $${budgetImpact} requested`,
-      fullLink,
-      btnText
-    );
-    if (importantNotification) msgs.push(importantNotification);
-  }
-
-  return msgs;
-};
-
-export const sendAndGetSlackCRNotifications = async (
-  teams: Team[],
-  changeRequest: Change_Request,
-  submitter: User,
-  wbsElement: WBS_Element,
-  projectWbsName: string
-) => {
-  const notifications: { channelId: string; ts: string }[] = [];
-  let message = '';
-  switch (changeRequest.type) {
-    case 'ACTIVATION':
-      message = `${submitter.firstName} ${submitter.lastName} wants to activate ${wbsElement.name} in ${projectWbsName}`;
-      break;
-    case 'STAGE_GATE':
-      message = `${submitter.firstName} ${submitter.lastName} wants to stage gate ${wbsElement.name} in ${projectWbsName}`;
-      break;
-    default:
-      message = `${changeRequest.type} CR submitted by ${submitter.firstName} ${submitter.lastName} for the ${projectWbsName} project`;
-  }
-
-  const completion: Promise<void>[] = teams.map(async (team) => {
-    const sentNotifications: { channelId: string; ts: string }[] = await sendSlackChangeRequestNotification(
-      team,
-      message,
-      changeRequest.crId
-    );
-    if (sentNotifications) notifications.push(...sentNotifications);
-  });
-  await Promise.all(completion);
-
-  return notifications;
-};
-
-export const sendSlackCRReviewedNotification = async (slackId: string, crId: number) => {
-  if (process.env.NODE_ENV !== 'production') return; // don't send msgs unless in prod
-  const msgs = [];
-  const fullMsg = `:tada: Your Change Request was just reviewed! Click the link to view! :tada:`;
-  const fullLink = `https://finishlinebyner.com/cr/${crId}`;
-  const btnText = `View CR#${crId}`;
-  msgs.push(sendMessage(slackId, fullMsg, fullLink, btnText));
-
-  return Promise.all(msgs);
-};
-
-/**
- * Replies and reacts to slack messages with the new change request status
- *
- * @param threads the threads of cr slack notifications to reply/react to
- * @param crId the cr id
- * @param approved is the cr approved
- */
-export const sendSlackCRStatusToThread = async (
-  threads: {
-    messageInfoId: string;
-    channelId: string;
-    timestamp: string;
-    changeRequestId: number;
-  }[],
-  crId: number,
-  approved: boolean
-) => {
-  if (process.env.NODE_ENV !== 'production') return; // don't send msgs unless in prod
-  const fullMsg = `This Change Request was ${approved ? 'approved! :tada:' : 'denied.'} Click the link to view.`;
-  const fullLink = `https://finishlinebyner.com/cr/${crId}`;
-  const btnText = `View CR#${crId}`;
-  try {
-    if (threads && threads.length !== 0) {
-      const msgs = threads.map((thread) =>
-        replyToMessageInThread(thread.channelId, thread.timestamp, fullMsg, fullLink, btnText)
-      );
-      const reactions = threads.map((thread) =>
-        reactToMessage(thread.channelId, thread.timestamp, approved ? 'white_check_mark' : 'x')
-      );
-      await Promise.all([...msgs, ...reactions]);
-    }
-  } catch (err: unknown) {
-    if (err instanceof Error) {
-      throw new HttpException(500, `Failed to send slack notification: ${err.message}`);
-    }
-  }
-};
-
-/**
- * Adds the relevant slack notifications for a change request to the change request
- *
- * @param crId the change request to add the slack threads to
- * @param notifications the slack threads to add to the change request
- */
-export const addSlackThreadsToChangeRequest = async (crId: number, threads: { channelId: string; ts: string }[]) => {
-  const promises = threads.map((notification) =>
-    prisma.message_Info.create({
-      data: {
-        changeRequestId: crId,
-        channelId: notification.channelId,
-        timestamp: notification.ts
-      },
-      include: {
-        changeRequest: true
-      }
-    })
-  );
-  await Promise.all(promises);
-};
 
 /**
  * This function updates the start date of all the blockings (and nested blockings) of the initial given work package.
